+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # Revision history for patch
 
+## Unreleased
+
+* `PatchMapWithMove` supports moves with a patch.
+
+* `PatchDMapWithMove` supports moves with a patch.
+
 ## 0.0.7.0 - 2022-06-23
 
 * Use `commutative-semigroups` for `Commutative`, making `Additive` a